recruitment-api-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai==1.33.0
python-dotenv==1.0.0
httpx==0.25.2
ultralytics==8.0.196
opencv-python==4.8.1.78